spike: adds tutorial view
