+++ src/views/onboarding/types.ts
@@ -1,6 +1,9 @@
 import { Collection as ApiCollection } from 'lib/learn-client/types'
 import { EnrichedNavItem } from 'components/sidebar/types'
 import { SidebarSidecarLayoutProps } from 'layouts/sidebar-sidecar'
+import { SidebarProps } from 'components/sidebar/types'
+import { NextPreviousProps } from 'views/tutorial-view/components'
+import { TutorialData } from 'views/tutorial-view'
 
 export interface OnboardingCollectionViewProps {
 	metadata: {
@@ -11,5 +14,16 @@ export interface OnboardingCollectionViewProps {
 	layoutProps: {
 		breadcrumbLinks: SidebarSidecarLayoutProps['breadcrumbLinks']
 		sidebarSections: EnrichedNavItem[]
+	}
+}
+
+export interface OnboardingTutorialViewProps {
+	tutorial: TutorialData & {
+		nextPreviousData: NextPreviousProps
+	}
+	layoutProps: {
+		headings: SidebarSidecarLayoutProps['headings']
+		breadcrumbLinks: SidebarSidecarLayoutProps['breadcrumbLinks']
+		navLevels: SidebarProps[]
 	}
 }
